Move error definitions out of cmd module
