+++ src/engine/strat_engine/mod.rs
@@ -9,13 +9,14 @@ mod cmd;
 mod device;
 mod dm;
 mod engine;
+mod errors;
 mod names;
 mod pool;
 mod serde_structs;
 mod thinpool;
 
-pub use self::cmd::{CmdError, CmdErrorKind};
 pub use self::engine::StratEngine;
+pub use self::errors::{CmdError, CmdErrorKind};
 
 #[cfg(test)]
 mod tests;
